requirements.txt
spotipy==2.24.0
ytmusicapi==1.8.2
python-dotenv==1.0.1
pytest==8.3.4
